[Refactor] Move reConnect/update wallets to utils
